git-svn-id: https://svn.apache.org/repos/asf/syncope/trunk@1528107 13f79535-47bb-0310-9956-ffa450edef68

+++ core/src/main/java/org/apache/syncope/core/workflow/user/activiti/ActivitiUserWorkflowAdapter.java
@@ -368,7 +368,11 @@ public class ActivitiUserWorkflowAdapter extends AbstractUserWorkflowAdapter {
         doExecuteTask(user, "delete", null);
         userDAO.delete(user);
 
-        historyService.deleteHistoricProcessInstance(user.getWorkflowId());
+        if (!historyService.createHistoricProcessInstanceQuery().
+                processInstanceId(user.getWorkflowId()).list().isEmpty()) {
+
+            historyService.deleteHistoricProcessInstance(user.getWorkflowId());
+        }
     }
 
     @Override
